refactor: skeleton util
move to util folder
add suport for predefined append targets
require append target to be specified
