+++ frontend/src/ts/elements/leaderboards.ts
@@ -8,7 +8,7 @@ import { isAuthenticated } from "../firebase";
 import differenceInSeconds from "date-fns/differenceInSeconds";
 import { getHTMLById as getBadgeHTMLbyId } from "../controllers/badge-controller";
 import * as ConnectionState from "../states/connection";
-import * as Skeleton from "../popups/skeleton";
+import * as Skeleton from "../utils/skeleton";
 import { debounce } from "throttle-debounce";
 import Format from "../utils/format";
 import SlimSelect from "slim-select";
@@ -647,7 +647,7 @@ export function show(): void {
     Notifications.add("You can't view leaderboards while offline", 0);
     return;
   }
-  Skeleton.append(wrapperId);
+  Skeleton.append(wrapperId, "popups");
   if (!Misc.isPopupVisible("leaderboardsWrapper")) {
     if (isAuthenticated()) {
       $("#leaderboardsWrapper #leaderboards .rightTableJumpToMe").removeClass(
