+++ frontend/src/ts/pages/about.ts
@@ -5,7 +5,7 @@ import * as Notifications from "../elements/notifications";
 import * as ChartController from "../controllers/chart-controller";
 import * as ConnectionState from "../states/connection";
 import intervalToDuration from "date-fns/intervalToDuration";
-import * as Skeleton from "../popups/skeleton";
+import * as Skeleton from "../utils/skeleton";
 
 function reset(): void {
   $(".pageAbout .contributors").empty();
