+++ frontend/src/ts/pages/loading.ts
@@ -1,6 +1,6 @@
 import * as Misc from "../utils/misc";
 import Page from "./page";
-import * as Skeleton from "../popups/skeleton";
+import * as Skeleton from "../utils/skeleton";
 
 export function updateBar(percentage: number, fast = false): void {
   const speed = fast ? 100 : 1000;
@@ -50,7 +50,7 @@ export const page = new Page(
     Skeleton.remove("pageLoading");
   },
   async () => {
-    Skeleton.append("pageLoading");
+    Skeleton.append("pageLoading", "main");
   },
   async () => {
     //
