+++ frontend/src/ts/pages/settings.ts
@@ -20,7 +20,7 @@ import { areFunboxesCompatible } from "../test/funbox/funbox-validation";
 import { get as getTypingSpeedUnit } from "../utils/typing-speed-units";
 import SlimSelect from "slim-select";
 
-import * as Skeleton from "../popups/skeleton";
+import * as Skeleton from "../utils/skeleton";
 import * as CustomBackgroundFilter from "../elements/custom-background-filter";
 
 type SettingsGroups<T extends SharedTypes.ConfigValue> = Record<
