+++ frontend/src/ts/popups/animated-modal.ts
@@ -1,5 +1,5 @@
 import { isPopupVisible } from "../utils/misc";
-import * as Skeleton from "./skeleton";
+import * as Skeleton from "../utils/skeleton";
 
 type CustomAnimation = {
   from: Record<string, string>;
@@ -33,6 +33,7 @@ export default class AnimatedModal {
   private wrapperId: string;
   private open = false;
   private setupRan = false;
+  private skeletonAppendParent: Skeleton.SkeletonAppendParents;
   private customShowAnimations: CustomWrapperAndModalAnimations | undefined;
   private customHideAnimations: CustomWrapperAndModalAnimations | undefined;
 
@@ -42,6 +43,7 @@ export default class AnimatedModal {
 
   constructor(
     wrapperId: string,
+    appendTo: Skeleton.SkeletonAppendParents,
     customAnimations?: ConstructorCustomAnimations,
     functions?: {
       customEscapeHandler?: (e: KeyboardEvent) => void;
@@ -53,6 +55,7 @@ export default class AnimatedModal {
       wrapperId = wrapperId.slice(1);
     }
 
+    this.skeletonAppendParent = appendTo;
     const dialogElement = document.getElementById(wrapperId);
     const modalElement = document.querySelector(
       `#${wrapperId} > .modal`
@@ -130,7 +133,7 @@ export default class AnimatedModal {
   async show(options?: ShowHideOptions): Promise<void> {
     // eslint-disable-next-line no-async-promise-executor
     return new Promise(async (resolve) => {
-      Skeleton.append(this.wrapperId);
+      Skeleton.append(this.wrapperId, this.skeletonAppendParent);
 
       if (!this.setupRan) {
         this.runSetup();
