+++ frontend/src/ts/popups/custom-test-duration-popup.ts
@@ -2,7 +2,7 @@ import * as UpdateConfig from "../config";
 import * as ManualRestart from "../test/manual-restart-tracker";
 import * as TestLogic from "../test/test-logic";
 import * as Notifications from "../elements/notifications";
-import * as Skeleton from "./skeleton";
+import * as Skeleton from "../utils/skeleton";
 import { isPopupVisible } from "../utils/misc";
 
 const wrapperId = "customTestDurationPopupWrapper";
@@ -72,7 +72,7 @@ function previewDuration(): void {
 }
 
 export function show(): void {
-  Skeleton.append(wrapperId);
+  Skeleton.append(wrapperId, "popups");
   if (!isPopupVisible(wrapperId)) {
     $("#customTestDurationPopupWrapper")
       .stop(true, true)
