+++ frontend/src/ts/popups/edit-tags-popup.ts
@@ -4,7 +4,7 @@ import * as Notifications from "../elements/notifications";
 import * as Loader from "../elements/loader";
 import * as Settings from "../pages/settings";
 import * as ConnectionState from "../states/connection";
-import * as Skeleton from "./skeleton";
+import * as Skeleton from "../utils/skeleton";
 import { isPopupVisible } from "../utils/misc";
 
 const wrapperId = "tagsWrapper";
@@ -17,7 +17,7 @@ export function show(action: string, id?: string, name?: string): void {
     return;
   }
 
-  Skeleton.append(wrapperId);
+  Skeleton.append(wrapperId, "popups");
 
   if (action === "add") {
     $("#tagsWrapper #tagsEdit").attr("action", "add");
