+++ frontend/src/ts/popups/result-tags-popup.ts
@@ -4,7 +4,7 @@ import * as Loader from "../elements/loader";
 import * as Notifications from "../elements/notifications";
 import * as ConnectionState from "../states/connection";
 import { areUnsortedArraysEqual, isPopupVisible } from "../utils/misc";
-import * as Skeleton from "./skeleton";
+import * as Skeleton from "../utils/skeleton";
 
 const wrapperId = "resultEditTagsPanelWrapper";
 
@@ -21,7 +21,7 @@ function show(): void {
     });
     return;
   }
-  Skeleton.append(wrapperId);
+  Skeleton.append(wrapperId, "popups");
   updateButtons();
   if (!isPopupVisible(wrapperId)) {
     $("#resultEditTagsPanelWrapper")
