+++ frontend/src/ts/test/practise-words.ts
@@ -5,7 +5,7 @@ import * as CustomText from "./custom-text";
 import * as TestInput from "./test-input";
 import * as ConfigEvent from "../observables/config-event";
 import { setCustomTextName } from "../states/custom-text-name";
-import * as Skeleton from "../popups/skeleton";
+import * as Skeleton from "../utils/skeleton";
 import { isPopupVisible } from "../utils/misc";
 
 const wrapperId = "practiseWordsPopupWrapper";
@@ -147,7 +147,7 @@ export function showPopup(): void {
     Notifications.add("Practice words is unsupported in zen mode", 0);
     return;
   }
-  Skeleton.append(wrapperId);
+  Skeleton.append(wrapperId, "popups");
   if (!isPopupVisible(wrapperId)) {
     $("#practiseWordsPopupWrapper")
       .stop(true, true)
